restorane3.0/requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]